+++ functions.php
@@ -32,12 +32,17 @@ require_once locate_template('/assets/jetpack.php' );
 require_once locate_template('/assets/custom-header.php');
 // Custom functions that cleaning unuse element of HTML, CSS, WordPress.
 require_once locate_template('/assets/cleanup.php');
+// Content related functions to manipulation your post.
+require_once locate_template('/assets/content.php');
+// Post Navigation Control: Breadcrumb, Pagination.
+require_once locate_template('/assets/post-navigation.php');
 
 if (!isset( $content_width))
     $content_width = 640;
 /**
  * Adjusts content_width value for full-width and single image attachment
  * templates, and when there are no active widgets in the sidebar.
+ *
  * @since NARGA v1.3.3
  * @from Twenty Twelve
  */
@@ -94,6 +99,8 @@ add_action('after_setup_theme', 'narga_setup');
 
 /**
  * Enqueue Scripts and Styles for Front-End
+ *
+ * @since NARGA v0.1
  **/
 function narga_assets() {
     global $wp_styles;
@@ -101,13 +108,13 @@ function narga_assets() {
     if ( !is_admin() ) {
 
         # Loads Foundation Main stylesheet
-        wp_enqueue_style( 'foundation', get_template_directory_uri() . '/stylesheets/foundation.min.css', array(), '2013-08-12', false );
+        wp_enqueue_style( 'foundation', get_template_directory_uri() . '/stylesheets/foundation.min.css', array(), '2013-08-12', 'all' );
 
         # Load Google Fonts API
-        wp_enqueue_style( 'google-font',"http://fonts.googleapis.com/css?family=Oswald|Open+Sans:400,400italic,700,700italic", array(), '2013-08-12', false );
+        wp_enqueue_style( 'google-font',"http://fonts.googleapis.com/css?family=Oswald|Open+Sans:400,400italic,700,700italic", array(), '2013-08-12', 'all' );
 
         # Loads our main stylesheet.
-        wp_enqueue_style( 'narga-style', get_stylesheet_uri(), array(), '2013-08-12'  );
+        wp_enqueue_style( 'narga-style', get_stylesheet_uri(), array(), '2013-08-12', 'all' );
 
         # Load JavaScripts
         wp_enqueue_script( 'modernizr', get_template_directory_uri() . '/javascripts/vendor/custom.modernizr.js', array( 'jquery' ), '2.6.2', true );
@@ -124,24 +131,8 @@ function narga_assets() {
 add_action( 'wp_enqueue_scripts', 'narga_assets' );
 
 /**
- * Enqueue Zepto to footer
- * Idea by ZGani 
- * @since NARGA v1.4.0
- **/
-if (!function_exists('narga_enqueue_zepto')) :
-    function narga_enqueue_zepto(){
-        echo '<!-- Check for Zepto support, load jQuery if necessary -->
-            <script type="text/javascript">
-document.write(\'<script src=' . get_template_directory_uri() . '/javascripts/vendor/\'
-+ (\'__proto__\' in {} ? \'zepto\' : \'jquery\')
-+ \'.min.js><\/script>\');
-        </script>';
-    } 
-  add_action('wp_footer', 'narga_enqueue_zepto');
-endif;
-
-/**
  * Includes the pro and custom functions if it exists
+ *
  * @since NARGA v1.1
  **/
 locate_template( array( 'assets/pro-functions.php', 'assets/custom-functions.php' ), true, false);
@@ -163,134 +154,9 @@ function narga_add_custom_css_file() {
 }
 
 /**
- * Creates a nicely formatted and more specific title element text
- * for output in head of document, based on current view.
- *
- * based on Twenty Twelve 1.0
- *
- * @param string $title Default title text for current view.
- * @param string $sep Optional separator.
- * @return string Filtered title.
- */
-if (!function_exists('narga_wp_title')) :  
-    function narga_wp_title( $title, $sep ) {
-        global $paged, $page;
-
-        if ( is_feed() )
-            return $title;
-
-        // Add the site name.
-        $title .= get_bloginfo( 'name' );
-
-        // Add the site description for the home/front page.
-        $site_description = get_bloginfo( 'description', 'display' );
-        if ( $site_description && ( is_home() || is_front_page() ) )
-            $title = "$title $sep $site_description";
-
-        // Add a page number if necessary.
-        if ( $paged >= 2 || $page >= 2 )
-            $title = "$title $sep " . sprintf( __( 'Page %s', 'narga' ), max( $paged, $page ) );
-
-        return $title;
-    }
-add_filter( 'wp_title', 'narga_wp_title', 10, 2 );
-endif;
-
-/**
- * Return entry meta information for posts, used by multiple loops
- * @since NARGA v1.1
- */
-if (!function_exists('narga_entry_meta')) :  
-    function narga_entry_meta() {
-        echo '<p class="post-meta-data">';
-        if (comments_open()) :
-            echo ' <span class="entry-comments right">';
-        comments_popup_link( __( 'No comment', 'narga' ), __( '1 comment', 'narga'),  __( '% comments', 'narga' ),  __( 'comments-link', 'narga' ),  __( 'Comments are off for this post', 'narga' ));
-        echo '</span>';
-        endif;
-        echo '<a href="' . get_permalink() . '" title="' . get_the_time() . '" rel="bookmark"><time class="updated" datetime="'. get_the_time('c') .'">'. sprintf(__('%s', 'narga'), get_the_time('M jS, Y'), get_the_time()) .'</time></a>';
-        if (false === get_post_format()) {
-            echo __(' in ', 'narga') .'<span class="entry-categories">' . get_the_category_list( ', ' ) . '</span> <span class="byline author">' . __(' by ', 'narga') . '<a href="'. get_author_posts_url(get_the_author_meta('ID')) .'" rel="author" class="fn">'. get_the_author() .'</a></span>';            
-        } else {
-            echo '';
-        }
-        edit_post_link(__('Edit', 'narga'), ' | ', '');
-        echo '</p>';
-    }
-endif;
-
-/**
- * Replace Read more link text
- * @since NARGA v1.6
- */
-if (!function_exists('narga_more_link')) :  
-    function narga_more_link( $more_link, $more_link_text ) {
-        $readmore = narga_options(post_readmore);
-        return str_replace( $more_link_text, $readmore, $more_link );
-    }
-    add_filter( 'the_content_more_link', 'narga_more_link', 10, 2 );
-endif;
-
-/**
- * WordPress Comments Adjustment
- *
- * @since NARGA v1.1
- **/
-
-if (!function_exists('narga_comments')) :  
-    function narga_comments($comment, $args, $depth) {
-        $GLOBALS['comment'] = $comment;
-        switch ( $comment->comment_type ) :
-                case 'pingback' :
-                case 'trackback' :
-                // Display trackbacks differently than normal comments.
-        ?>
-        <li <?php comment_class(); ?> id="comment-<?php comment_ID(); ?>">
-                <p><?php _e( 'Pingback:', 'narga' ); ?> <?php comment_author_link(); ?> <?php edit_comment_link( __( '(Edit)', 'narga' ), '<span class="edit-link">', '</span>' ); ?></p>
-<?php
-                    break;
-                default :
-                    // Proceed with normal comments.
-                    global $post;
-                    echo '<li ';
-                    comment_class();
-                    echo '>
-                        <article id="comment-' . get_comment_ID() . '" class="comment">
-                        <header>
-                        <div class="comment-author vcard">';
-                    echo get_avatar($comment,64);
-                    printf(__('<cite class="fn">%s</cite>', 'narga'), get_comment_author_link(),
-                        // If current post author is also comment author, make it known visually.
-                        ( $comment->user_id === $post->post_author ) ? '<span> ' . __( 'Post author', 'narga' ) . '</span>' : '');
-                    echo '<div class="comment-meta">
-                        <time datetime="' . get_comment_date('c') . '"  itemprop="commentTime"><a itemprop="url" href="' . htmlspecialchars( get_comment_link( $comment->comment_ID ) ) . '">';
-                    printf(__('%1$s', 'narga'), get_comment_date(),  get_comment_time());
-                    echo '</a></time>
-                        </div>
-                        </div>
-                        </header>
-                        <section itemprop="commentText" class="comment">';
-                    if ($comment->comment_approved == '0') : 
-                        echo '<div class="comment-awaiting-moderation">';
-                    _e('Your comment is awaiting moderation.', 'narga');
-                    echo '</div>';
-endif;
-comment_text();
-$id = $comment->comment_ID;
-edit_comment_link(__('Edit', 'narga'), '<a class="comment-del-link" href="'.admin_url("comment.php?action=cdc&c=$id").'">' . __( 'Del', 'narga' ) . '</a> ', '<a class="comment-spam-link" href="'.admin_url("comment.php?action=cdc&dt=spam&c=$id").'">' . __( 'Spam', 'narga' ) . '</a>');
-echo '<span class="reply">';
-comment_reply_link( array_merge( $args, array( 'reply_text' => __( 'Reply', 'narga' ), 'after' => '<span>&darr;</span>', 'depth' => $depth, 'max_depth' => $args['max_depth'] ) ) );
-echo '</span>
-    </section>
-    </article>';
-break;
-endswitch; // end comment_type check
-    }
-endif;
-
-/**
  * Orbit Slider as Featured Post
  * Function to render orbit slide based on featured category and number of slide in Customize.
+ *
  * @since NARGA v1.1
  */
 if (!function_exists('narga_orbit_slider')) :  
@@ -341,110 +207,6 @@ if (narga_options('slide_indicator') == 1) :
 endif;
 echo '</div>';
     }
-endif;
-
-/**
- * Rewrite default wordpress pagination function
- * Since NARGA v1.1
- **/
-if (!function_exists('narga_pagination')) :  
-    function narga_pagination() {
-        global $wp_query;
-        $big = 999999999; # This needs to be an unlikely integer
-        $paginate_links = paginate_links( array(
-            'base' => str_replace( $big, '%#%', get_pagenum_link($big) ),
-            'current' => max( 1, get_query_var('paged') ),
-            'total' => $wp_query->max_num_pages,
-            'mid_size' => 5,
-            'prev_next' => true,
-            'prev_text' => __('&laquo; Previous', 'narga'),
-            'next_text' => __('Next &raquo;', 'narga'),
-            'type' => 'list'
-        ) );
-        $search  = array('ul class=\'page-numbers', '<li><a class="next page-numbers"');
-        $replace = array('ul class=\'pagination', '<li class="arrow"><a');
-        echo '<nav id="post-nav" class="columns">' .str_replace($search, $replace, $paginate_links) . '</nav>';
-    }
-endif;
-
-/**
- * Rewrite default wordpress comments pagination function
- * Since NARGA v1.1
- **/
-if (!function_exists('narga_comment_pagination')) :  
-    function narga_comment_pagination() {
-        //read the page links but do not echo
-        $comment_page = paginate_comments_links( array(
-            'prev_text' => __('&laquo; Previous', 'narga'),
-            'next_text' => __('Next &raquo;', 'narga'),
-            'echo' => false, 
-            'type' => 'list'
-        ) );
-        $search  = array('ul class=\'page-numbers', '<li><a class="next page-numbers"');
-        $replace = array('ul class=\'pagination', '<li class="arrow"><a');
-        echo '<nav id="comments-nav">' . str_replace($search, $replace,  $comment_page) . '</nav>';
-    }
-endif;
-
-/**
- * Generator breadcrumbs with Foundation
- * Since NARGA v1.3.5
- **/
-if (!function_exists('narga_breadcrumb')) :  
-    function narga_breadcrumb() {
-        if (!is_home()) {
-            echo '<ul class="breadcrumbs"><li><a href="';
-            echo home_url();
-            echo '">';
-            bloginfo('name');
-            echo "</a></li>";
-            if (is_category()) {
-                the_category('<li>', '</li>');
-            } elseif (is_page() || is_single()) {
-                echo the_title('<li class="current">', '</li>');
-            } 
-            echo "</ul>";
-        }
-    }
-endif;
-
-
-# Function to trim the excerpt
-if (!function_exists('narga_excerpts')) :
-    function narga_excerpts($text = false) {
-        # If is the home page, an archive, or search results
-        if(!is_singular()) :
-            global $post;
-            $excerpt_length = narga_options('excerpt_length');
-            $text = $post->post_excerpt;
-            # If an excerpt is set in the Optional Excerpt box
-            if ( empty($post->post_excerpt) ) {
-                $content = $post->post_content;
-                if (false == get_post_format() || count($content) > $excerpt_length) {
-                    $content = str_replace('\]\]\>', ']]&gt;', $content);
-                    $content = preg_replace('@<script[^>]*?>.*?</script>@si', '', $content);
-                    $content = strip_shortcodes($content);
-                    $content = strip_tags($content, '<p>');
-                    $words = explode(' ', $content, $excerpt_length + 1);
-                    if (count($words)> $excerpt_length) {
-                        array_pop($words);
-                        array_push($words, '...<br><a href="'.get_permalink($post->ID) .'" class="more-link">' . __('Continue Reading »', 'narga') . '</a>');
-                        $text = wpautop(implode(' ', $words));
-                    }
-                }
-            } else {
-                $text = apply_filters('the_excerpt', $text);
-            }
-            endif;
-            # Make sure to return the content
-            return $text;
-    }
-
-# Replace content with excerpt
-if (narga_options('excerpt_length') != '0') :
-    remove_filter('get_the_excerpt', 'wp_trim_excerpt');
-    add_filter('the_content', 'narga_excerpts');
-endif;
 endif;
 
 /**